# Refactor and optimize BlockedModules