--- a/QRCoder/QRCodeGenerator.ModulePlacer.BlockedModules.cs
+++ b/QRCoder/QRCodeGenerator.ModulePlacer.BlockedModules.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections;
+﻿using System;
+using System.Collections;
 
 namespace QRCoder
 {
@@ -9,9 +10,12 @@ namespace QRCoder
             /// <summary>
             /// Struct that represents blocked modules using rectangles.
             /// </summary>
-            public struct BlockedModules
+            public struct BlockedModules : IDisposable
             {
-                private readonly BitArray[] _blockedModules;
+                private readonly bool[,] _blockedModules;
+
+                [ThreadStatic]
+                private static bool[,] _staticBlockedModules;
 
                 /// <summary>
                 /// Initializes a new instance of the <see cref="BlockedModules"/> struct with a specified capacity.
@@ -19,10 +23,19 @@ namespace QRCoder
                 /// <param name="capacity">The initial capacity of the blocked modules list.</param>
                 public BlockedModules(int size)
                 {
-                    _blockedModules = new BitArray[size];
-                    for (int i = 0; i < size; i++)
+                    if (_staticBlockedModules != null && _staticBlockedModules.GetUpperBound(0) >= (size - 1))
                     {
-                        _blockedModules[i] = new BitArray(size);
+                        _blockedModules = _staticBlockedModules;
+                        _staticBlockedModules = null;
+#if NET6_0_OR_GREATER
+                        Array.Clear(_blockedModules);
+#else
+                        Array.Clear(_blockedModules, 0, _blockedModules.Length);
+#endif
+                    }
+                    else
+                    {
+                        _blockedModules = new bool[size, size];
                     }
                 }
 
@@ -33,7 +46,7 @@ namespace QRCoder
                 /// <param name="y">The y-coordinate of the module.</param>
                 public void Add(int x, int y)
                 {
-                    _blockedModules[y][x] = true;
+                    _blockedModules[y, x] = true;
                 }
 
                 /// <summary>
@@ -46,7 +59,7 @@ namespace QRCoder
                     {
                         for (int x = rect.X; x < rect.X + rect.Width; x++)
                         {
-                            _blockedModules[y][x] = true;
+                            _blockedModules[y, x] = true;
                         }
                     }
                 }
@@ -59,7 +72,7 @@ namespace QRCoder
                 /// <returns><c>true</c> if the coordinates are blocked; otherwise, <c>false</c>.</returns>
                 public bool IsBlocked(int x, int y)
                 {
-                    return _blockedModules[y][x];
+                    return _blockedModules[y, x];
                 }
 
                 /// <summary>
@@ -73,11 +86,19 @@ namespace QRCoder
                     {
                         for (int x = r1.X; x < r1.X + r1.Width; x++)
                         {
-                            if (_blockedModules[y][x])
+                            if (_blockedModules[y, x])
                                 return true;
                         }
                     }
                     return false;
+                }
+
+                public void Dispose()
+                {
+                    if (_staticBlockedModules == null || _staticBlockedModules.GetUpperBound(0) < _blockedModules.GetUpperBound(0))
+                    {
+                        _staticBlockedModules = _blockedModules;
+                    }
                 }
             }
         }