--- a/QRCoderApiTests/net35+net40+net50+net50-windows+netstandard20/QRCoder.approved.txt
+++ b/QRCoderApiTests/net35+net40+net50+net50-windows+netstandard20/QRCoder.approved.txt
@@ -868,6 +868,7 @@ namespace QRCoder
     {
         public QRCodeData(int version) { }
         public QRCodeData(byte[] rawData, QRCoder.QRCodeData.Compression compressMode) { }
+        public QRCodeData(int version, bool addPadding) { }
         public QRCodeData(string pathToRawData, QRCoder.QRCodeData.Compression compressMode) { }
         public System.Collections.Generic.List<System.Collections.BitArray> ModuleMatrix { get; set; }
         public int Version { get; }