--- a/QRCoderApiTests/net60-windows/QRCoder.approved.txt
+++ b/QRCoderApiTests/net60-windows/QRCoder.approved.txt
@@ -876,6 +876,7 @@ namespace QRCoder
     {
         public QRCodeData(int version) { }
         public QRCodeData(byte[] rawData, QRCoder.QRCodeData.Compression compressMode) { }
+        public QRCodeData(int version, bool addPadding) { }
         public QRCodeData(string pathToRawData, QRCoder.QRCodeData.Compression compressMode) { }
         public System.Collections.Generic.List<System.Collections.BitArray> ModuleMatrix { get; set; }
         public int Version { get; }