--- a/QRCoderApiTests/netstandard13/QRCoder.approved.txt
+++ b/QRCoderApiTests/netstandard13/QRCoder.approved.txt
@@ -775,6 +775,7 @@ namespace QRCoder
     {
         public QRCodeData(int version) { }
         public QRCodeData(byte[] rawData, QRCoder.QRCodeData.Compression compressMode) { }
+        public QRCodeData(int version, bool addPadding) { }
         public QRCodeData(string pathToRawData, QRCoder.QRCodeData.Compression compressMode) { }
         public System.Collections.Generic.List<System.Collections.BitArray> ModuleMatrix { get; set; }
         public int Version { get; }