@codebude 2024-05-19
![image](https://github.com/codebude/QRCoder/assets/2236234/66ac1343-59c9-4483-834f-cc3861d622b3)

@Shane32 I merged the other PRs as suggested, but it seems there's still a conflict. Can you pull the current master into your branch?

@gfoidl Another performance optimization. Would you mind doing a review?